RequirementMachine: Split off ProtocolGraph into its own file
