+++ include/swift/AST/LayoutConstraintKind.h
@@ -14,6 +14,8 @@
 //
 //===----------------------------------------------------------------------===//
 
+#include "llvm/Support/DataTypes.h"
+
 #ifndef SWIFT_LAYOUT_CONSTRAINTKIND_H
 #define SWIFT_LAYOUT_CONSTRAINTKIND_H
