+++ lib/AST/RequirementMachine.cpp
@@ -15,6 +15,7 @@
 #include "swift/AST/Decl.h"
 #include "swift/AST/GenericSignature.h"
 #include "swift/AST/PrettyStackTrace.h"
+#include "swift/AST/ProtocolGraph.h"
 #include "swift/AST/Requirement.h"
 #include "swift/AST/RewriteSystem.h"
 #include "llvm/ADT/DenseSet.h"
@@ -25,161 +26,6 @@ using namespace swift;
 using namespace rewriting;
 
 namespace {
-
-struct ProtocolInfo {
-  ArrayRef<ProtocolDecl *> Inherited;
-  llvm::TinyPtrVector<AssociatedTypeDecl *> AssociatedTypes;
-  ArrayRef<Requirement> Requirements;
-
-  // Used by computeDepth() to detect circularity.
-  unsigned Mark : 1;
-
-  // Longest chain of protocol refinements, including this one.
-  // Greater than zero on valid code, might be zero if there's
-  // a cycle.
-  unsigned Depth : 31;
-
-  // Index of the protocol in the linear order.
-  unsigned Index : 32;
-
-  ProtocolInfo() {
-    Mark = 0;
-    Depth = 0;
-    Index = 0;
-  }
-
-  ProtocolInfo(ArrayRef<ProtocolDecl *> inherited,
-               llvm::TinyPtrVector<AssociatedTypeDecl *> &&types,
-               ArrayRef<Requirement> reqs)
-    : Inherited(inherited),
-      AssociatedTypes(types),
-      Requirements(reqs) {
-    Mark = 0;
-    Depth = 0;
-    Index = 0;
-  }
-};
-
-struct ProtocolGraph {
-  llvm::DenseMap<const ProtocolDecl *, ProtocolInfo> Info;
-  std::vector<const ProtocolDecl *> Protocols;
-  bool Debug = false;
-
-  void visitRequirements(ArrayRef<Requirement> reqs) {
-    for (auto req : reqs) {
-      if (req.getKind() == RequirementKind::Conformance) {
-        addProtocol(req.getProtocolDecl());
-      }
-    }
-  }
-
-  const ProtocolInfo &getProtocolInfo(
-      const ProtocolDecl *proto) const {
-    auto found = Info.find(proto);
-    assert(found != Info.end());
-    return found->second;
-  }
-
-  void addProtocol(const ProtocolDecl *proto) {
-    if (Info.count(proto) > 0)
-      return;
-
-    Info[proto] = {proto->getInheritedProtocols(),
-                   proto->getAssociatedTypeMembers(),
-                   proto->getRequirementSignature()};
-    Protocols.push_back(proto);
-  }
-
-  void computeTransitiveClosure() {
-    unsigned i = 0;
-    while (i < Protocols.size()) {
-      auto *proto = Protocols[i++];
-      visitRequirements(getProtocolInfo(proto).Requirements);
-    }
-  }
-
-  void computeLinearOrder() {
-    for (const auto *proto : Protocols) {
-      (void) computeProtocolDepth(proto);
-    }
-
-    std::sort(
-        Protocols.begin(), Protocols.end(),
-        [&](const ProtocolDecl *lhs,
-            const ProtocolDecl *rhs) -> bool {
-          const auto &lhsInfo = getProtocolInfo(lhs);
-          const auto &rhsInfo = getProtocolInfo(rhs);
-
-          // protocol Base {} // depth 1
-          // protocol Derived : Base {} // depth 2
-          //
-          // Derived < Base in the linear order.
-          if (lhsInfo.Depth != rhsInfo.Depth)
-            return lhsInfo.Depth > rhsInfo.Depth;
-
-          return TypeDecl::compare(lhs, rhs) < 0;
-        });
-
-    for (unsigned i : indices(Protocols)) {
-      Info[Protocols[i]].Index = i;
-    }
-
-    if (Debug) {
-      for (const auto *proto : Protocols) {
-        const auto &info = getProtocolInfo(proto);
-        llvm::dbgs() << "@ Protocol " << proto->getName()
-                     << " Depth=" << info.Depth
-                     << " Index=" << info.Index << "\n";
-      }
-    }
-  }
-
-  void computeInheritedAssociatedTypes() {
-    for (const auto *proto : Protocols) {
-      auto &info = Info[proto];
-
-      llvm::SmallDenseSet<const AssociatedTypeDecl *, 4> visited;
-      for (const auto *inherited : info.Inherited) {
-        if (inherited == proto)
-          continue;
-
-        for (auto *inheritedType : getProtocolInfo(inherited).AssociatedTypes) {
-          if (!visited.insert(inheritedType).second)
-            continue;
-
-          // The 'if (inherited == proto)' above avoids a potential
-          // iterator invalidation here.
-          info.AssociatedTypes.push_back(inheritedType);
-        }
-      }
-    }
-  }
-
-private:
-  unsigned computeProtocolDepth(const ProtocolDecl *proto) {
-    auto &info = Info[proto];
-
-    if (info.Mark) {
-      // Already computed, or we have a cycle. Cycles are diagnosed
-      // elsewhere in the type checker, so we don't have to do
-      // anything here.
-      return info.Depth;
-    }
-
-    info.Mark = true;
-    unsigned depth = 0;
-
-    for (auto *inherited : info.Inherited) {
-      unsigned inheritedDepth = computeProtocolDepth(inherited);
-      depth = std::max(inheritedDepth, depth);
-    }
-
-    depth++;
-
-    info.Depth = depth;
-    return depth;
-  }
-};
 
 struct RewriteSystemBuilder {
   ASTContext &Context;
